% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netaipq-package.R
\docType{package}
\name{netaipq-package}
\alias{netaipq}
\alias{netaipq-package}
\title{netaipq: Networks and Machine Learning for 'phyloseq' Objects}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Network, graph, causal-inference, and machine-learning tools for 'phyloseq' objects within the 'pqverse' ecosystem. Provides genuinely new analysis methods (co-occurrence and association networks, bipartite-network visualisation, supervised classification, indicator-species and differential-abundance analyses) that go beyond simple 'ggplot2' wrappers. 'netaipq' is the home for any item adding a brand-new machine-learning, network, graph, or statistical-analysis dependency, keeping 'MiscMetabar' lean. Built on top of 'MiscMetabar'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adrientaudiere.github.io/netaipq/}
  \item \url{https://github.com/adrientaudiere/netaipq}
  \item Report bugs at \url{https://github.com/adrientaudiere/netaipq/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]
}

}
\keyword{internal}
